py_src/requirements.txt
aiogram
aiohttp
python-dotenv
aiogram_calendar
apscheduler

